refactor to move extended lookup table value set class to fhirengine, moved tests
