+++ prime-router/src/main/kotlin/fhirengine/engine/CustomFhirPathFunctions.kt
@@ -145,50 +145,4 @@ class CustomFhirPathFunctions : FhirPathFunctions {
             mutableListOf(StringType(result))
         }
     }
-
-    /**
-     * Get a list of value/key pairs from the specified lookup table and columns
-     * @return a map with the value/key pairs from the specified table and columns
-     */
-//    fun valueSetTableLookup(
-//        focus: MutableList<Base>,
-//        parameters: MutableList<MutableList<Base>>?,
-//        metadata: Metadata = Metadata.getInstance()
-//    ): SortedMap<String, String> {
-//        if (parameters.isNullOrEmpty() || parameters.size != 3) {
-//            throw SchemaException("Must provide lookup table name, key column, and value column")
-//        }
-//
-//        val tableName = parameters[0].toString()
-//        if (tableName.isNullOrBlank()) {
-//            throw SchemaException("No lookup table name specified")
-//        }
-//
-//        val keyColumn = parameters[1].toString()
-//        if (keyColumn.isNullOrBlank()) {
-//            throw SchemaException("No key column name specified")
-//        }
-//
-//        val valueColumn = parameters[2].toString()
-//        if (valueColumn.isNullOrBlank()) {
-//            throw SchemaException("No value column name specified")
-//        }
-//
-//        val lookupTable = metadata.findLookupTable(name = tableName)
-//            ?: throw SchemaException("Specified lookup table not found")
-//
-//        if (!lookupTable.hasColumn(keyColumn)) {
-//            throw SchemaException("Key column not found in specified lookup table")
-//        }
-//
-//        if (!lookupTable.hasColumn(valueColumn)) {
-//            throw SchemaException("Value column not found in specified lookup table")
-//        }
-//
-//        val filterTable = lookupTable.table.retainColumns(keyColumn, valueColumn)
-//        // transform into sortedmap - filterTable.stream().forEach(row ->)
-//
-//        // stub return
-//        return sortedMapOf(Pair("A", "B"))
-//    }
 }
